requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
boto3==1.21.13
botocore==1.24.13
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
coverage==6.3.2
dateparser==1.1.0
dateutils==0.6.12
dj-database-url==0.5.0
Django==3.2.5
django-admin-searchable-dropdown==1.2
django-constance==2.8.0
django-heroku==0.3.1
django-picklefield==3.0.1
django-reverse-admin==2.9.6
django-storages==1.12.3
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
jmespath==0.10.0
oauthlib==3.2.0
protobuf==3.19.4
psycopg-binary==3.0.9
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-Levenshtein==0.12.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
thefuzz==0.19.0
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==6.0.0
